学生个人信息页面完善

+++ tutor-system/src/main/java/com/zty/system/service/impl/ParentServiceImpl.java
@@ -5,6 +5,7 @@ import com.zty.common.utils.DateUtils;
 import com.zty.common.utils.StringUtils;
 import com.zty.system.domain.Parent;
 import com.zty.system.domain.SysParent;
+import com.zty.system.domain.SysUserPost;
 import com.zty.system.domain.vo.ParentVo;
 import com.zty.system.mapper.*;
 import com.zty.system.service.IParentService;
@@ -104,9 +105,32 @@ public class ParentServiceImpl implements IParentService
     public int updateParent(Parent parent)
     {
         parent.setUpdateTime(DateUtils.getNowDate());
-        parentMapper.deleteSysParentByUserId(parent.getUserId());
-        insertSysParent(parent);
+        if (parent.getSysParent().getId()!=null){
+            parentMapper.deleteSysParentByUserId(parent.getUserId());
+            insertSysParent(parent);
+        }
+        // 删除用户与岗位关联
+        userPostMapper.deleteUserPostByUserId(parent.getUserId());
+        // 新增用户与岗位管理
+        insertUserPost(parent.getUserId(),parent.getPostIds());
         return parentMapper.updateParent(parent);
+    }
+
+    public void insertUserPost(Long userId,Long[] posts)
+    {
+        if (StringUtils.isNotEmpty(posts))
+        {
+            // 新增用户与岗位管理
+            List<SysUserPost> list = new ArrayList<SysUserPost>(posts.length);
+            for (Long postId : posts)
+            {
+                SysUserPost up = new SysUserPost();
+                up.setUserId(userId);
+                up.setPostId(postId);
+                list.add(up);
+            }
+            userPostMapper.batchUserPost(list);
+        }
     }
 
     /**
